requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.13.1
gevent==21.12.0
greenlet==1.1.2
hiredis==2.0.0
httpie==2.6.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.2.2
msgpack==1.0.3
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==20.0.1
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.1
redis==4.0.2
requests==2.26.0
requests-toolbelt==0.9.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
